+++ examples/python/20_leader_arm_state_check.py
@@ -26,7 +26,7 @@ import queue
 
 SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
 URDF_PATH = os.path.join(SCRIPT_DIR, "../../models/leader_arm", "model.urdf") # /../../models/leader_arm/model.urdf
-LEADER_ARM_DEVICE_NAME = rby.upc.MasterArmDeviceName
+LEADER_ARM_DEVICE_NAME = rby.upc.LeaderArmDeviceName
 
 class LeaderArm:
     DOF = 14